Clean up code
Modify code to conform to the community Ruby style guide more.

- Prefer the use of MatchData over $1.
- Align the parameters of a method call if they span more than one line.
- Use || instead of or.
- Remove assignment in condition.
- Avoid using {...} for multi-line blocks.
- Do not use :: for method calls.
- Use def with parentheses when there are arguments.
- Missing top-level class documentation comment.
- Missing top-level module documentation comment.
- Ruby 1.8 hash syntax detected.
- Favor modifier if/unless usage when you have a single-line body.
- Do not use parentheses for method calls with no arguments.
- Redundant `self` detected.
- Use %r for regular expressions matching more than 1 '/' character.
- Surrounding space missing for '{'.
- Prefer single-quoted strings when you don't need string interpolation.
- Trailing whitespace detected.
- Assigned but unused variable.
- Line is too long.

+++ spec/spec_helper.rb
@@ -71,9 +71,7 @@ Spork.each_run do
         screenshot_dir = Rails.root.join('tmp/screenshots').to_s
         unless screenshot.filename.blank?
           screenshot_file = "#{screenshot_dir}/#{screenshot.filename}"
-          if File.exists?(screenshot_file)
-            FileUtils.rm(screenshot_file)
-          end
+          FileUtils.rm(screenshot_file) if File.exists?(screenshot_file)
         end
       end
       DatabaseCleaner.clean
